% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.t.r
\name{aldex.robust.ttest}
\alias{aldex.robust.ttest}
\title{Calculate robustness of Welch's t-test statistics}
\usage{
aldex.robust.ttest(
  clr,
  paired.test = FALSE,
  verbose = FALSE,
  instances = c(2, 4, 8, 32, 64, 128)
)
}
\arguments{
\item{clr}{An \code{ALDEx2} object. The output of \code{aldex.clr}.}

\item{paired.test}{Toggles whether to calculate paired tests.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information
while running. Useful for debugging errors on large datasets. Applies
to \code{effect = TRUE}.}

\item{instances}{a vector of number of instances to test}
}
\value{
Returns a \code{data.frame} with the number of adjusted p values 
 below a FDR of 0.05 and the input number of instances vector
}
\description{
\code{aldex.robust.ttest} calculates the the posterior predictive 
 value of Welch's t-test on the data returned by \code{aldex.clr}.
}
\examples{
data(selex)
#subset for efficiency
selex <- selex[1201:1600,]
conds <- c(rep("NS", 7), rep("S", 7))
x <- aldex.clr(selex, conds, mc.samples=2, denom="all")
ttest.test <- aldex.ttest(x)
}
\references{
Please use the citation given by
 \code{citation(package="ALDEx2")}.
}
\seealso{
\code{\link{aldex}},
 \code{\link{aldex.clr}},
 \code{\link{aldex.ttest}},
 \code{\link{aldex.kw}},
 \code{\link{aldex.glm}},
 \code{\link{aldex.effect}},
 \code{\link{aldex.corr}},
 \code{\link{selex}}
}
\author{
Greg Gloor
}
