% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex-scaleSim.R
\name{interpretGamma}
\alias{interpretGamma}
\title{Interpret the scale model implied by a certain level of gamma}
\usage{
interpretGamma(clr)
}
\arguments{
\item{clr}{A aldex.clr object}
}
\value{
A table
}
\description{
Interpret the scale model implied by a certain level of gamma
}
