% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex-scaleSim.R
\name{plot_alpha}
\alias{plot_alpha}
\title{Create alpha diagram for scale simulation sensitivity result}
\usage{
plot_alpha(
  sen_results,
  test = "t",
  thresh = 0.05,
  taxa_to_label = 10,
  glmVar = NULL,
  bayesEst = TRUE
)
}
\arguments{
\item{sen_results}{A list return by aldex.senAnalysis()}

\item{test}{A character string. What test was used to calculate the results}

\item{thresh}{A numeric between 0 and 1. What threshold should be used for significance?}

\item{taxa_to_label}{A positive integer. How many taxa should be labeled in the plot?}

\item{glmVar}{If `test = "glm"`, what variable do you want plotted?}

\item{bayesEst}{A boolean. Do you want to use the Bayesian hypothesis testing method?}
}
\value{
A plot object
}
\description{
Create alpha diagram for scale simulation sensitivity result
}
