% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex-scaleSim.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Create gamma diagram for scale simulation sensitivity result}
\usage{
plot_gamma(
  sen_results,
  test = "t",
  thresh = 0.05,
  taxa_to_label = 10,
  glmVar = NULL,
  blackWhite = FALSE,
  cex = 1
)
}
\arguments{
\item{sen_results}{A list return by aldex.senAnalysis()}

\item{test}{A character string. What test was used to calculate the results}

\item{thresh}{A numeric between 0 and 1. What threshold should be used for significance?}

\item{taxa_to_label}{A positive integer. How many taxa should be labeled in the plot?}

\item{glmVar}{If `test = "glm"`, what variable do you want plotted?}

\item{blackWhite}{boolean. If TRUE, returns the plot in black and white.}

\item{cex}{Default == 1. Controls the size of the axis and text labels in the plots.}
}
\value{
A plot object
}
\description{
Create gamma diagram for scale simulation sensitivity result
}
