\name{selex}
\docType{data}
\alias{selex}
\title{Selection-based differential sequence variant abundance dataset}
\description{
  This data set gives the differential abundance of 1600 enzyme variants
  grown under selective (NS) and selective (S) conditions
}
\usage{data(selex)}
\format{A dataframe of 1600 features and 14 samples. The first 7 samples are non-selected,
 the last 7 are selected.}
\source{McMurrough et al (2014) PNAS doi:10.1073/pnas.1322352111}
\references{
 McMurrough et al (2014) PNAS doi:10.1073/pnas.1322352111
}
\keyword{datasets}
