\name{synth2}
\docType{data}
\alias{synth2}
\title{Synthetic asymmetric dataset}
\description{
  This synthetic dataset contains 2 percent sparsity as 0 values
  asymmetrically distributed. It is used as a test dataset.
}
\usage{selex}
\format{A dataframe of 1000 features and 16 samples. The first 8 samples contain
  20 features set to 0, the last 8 samples contain counts.}
\source{Gloor et al (2017) notes}
\keyword{datasets}
