% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_corr-tpq.R
\name{aldex.corr}
\alias{aldex.corr}
\title{Calculate correlation with a continuous variable}
\usage{
aldex.corr(clr, verbose = FALSE, ...)
}
\arguments{
\item{clr}{An \code{ALDEx2} object. The output of \code{aldex.clr}.}

\item{verbose}{A boolean. Toggles whether to print diagnostic information while
running. Useful for debugging errors on large datasets. Applies to
\code{effect = TRUE}.}

\item{...}{Arguments passed to \code{cor.test}.}
}
\value{
Returns a data.frame of the average
 coefficients and their p-values for each feature,
 with FDR appended as a \code{BH} column.
}
\description{
\code{aldex.corr} calculates the expected values for the correlation
 between each feature and a continuous variable, using data returned
 returned by \code{aldex.clr}.
}
\examples{
data(selex)
#subset for efficiency
selex <- selex[1201:1600,]
x <- aldex.clr(selex, 1:14)
corr.test <- aldex.corr(x)
}
\references{
Please use the citation given by
 \code{citation(package="ALDEx2")}.
}
\seealso{
\code{\link{aldex}},
 \code{\link{aldex.clr}},
 \code{\link{aldex.ttest}},
 \code{\link{aldex.kw}},
 \code{\link{aldex.glm}},
 \code{\link{aldex.effect}},
 \code{\link{aldex.corr}},
 \code{\link{selex}}
}
\author{
Thom Quinn
}
