% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_glm.r
\name{aldex.kw}
\alias{aldex.kw}
\title{Calculate the Kruskal-Wallis test and glm ANOVA statistics}
\usage{
aldex.kw(clr, conditions, useMC = FALSE)
}
\arguments{
\item{clr}{An \code{ALDEx2} object. The output of \code{aldex.clr}.}

\item{conditions}{A character vector. A description of the data structure used
for testing. Typically, a vector of group labels. For \code{aldex.glm}, use
a \code{model.matrix}.}

\item{useMC}{Toggles whether to use multi-core.}
}
\value{
Returns a \code{data.frame} with the following information:
\item{kw.ep}{ a vector containing the expected p-value of the Kruskal-Wallis test
 for each feature }
\item{kw.eBH}{ a vector containing the corresponding expected value of the
 Benjamini-Hochberg corrected p-value for each feature }
\item{glm.ep}{ a vector containing the expected p-value of the glm ANOVA
 for each feature }
\item{glm.eBH}{ a vector containing the corresponding expected value of the
 Benjamini-Hochberg corrected p-value for each feature }
}
\description{
\code{aldex.kw} calculates the expected values of the Kruskal-Wallis
 test and a glm ANOVA on the data returned by \code{aldex.clr}.
}
\examples{
data(selex)
#subset for efficiency
selex <- selex[1201:1600,]
conds <- c(rep("A", 4), rep("B", 3), rep("C", 7))
x <- aldex.clr(selex, conds, mc.samples=1, denom="all")
kw.test <- aldex.kw(x, conds)
}
\references{
Please use the citation given by
 \code{citation(package="ALDEx2")}.
}
\seealso{
\code{\link{aldex}},
 \code{\link{aldex.clr}},
 \code{\link{aldex.ttest}},
 \code{\link{aldex.kw}},
 \code{\link{aldex.glm}},
 \code{\link{aldex.effect}},
 \code{\link{aldex.corr}},
 \code{\link{selex}}
}
\author{
Arianne Albert
}
