% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScaleMatrix.r
\name{aldex.makeScaleMatrix}
\alias{aldex.makeScaleMatrix}
\title{Generate a differential scale matrix by group}
\usage{
aldex.makeScaleMatrix(gamma, diff, conditions, mc.samples = 128)
}
\arguments{
\item{gamma}{- the base gamma value for the sdlog parameter of rlnorm}

\item{diff}{- the difference in meanlog value, this is usually small}

\item{conditions}{- the conditions vector for the dataset}

\item{mc.samples}{- the number of Monte-Carlo instances used by aldex()}
}
\value{
returns a matrix of gamma values that are used as an estimate
of the posterior value of the geometric mean of the Monte-Carlo instances.
This allows different gamma values to be applied to each group and has the
result of moving the centre of mass of the data towards the midline of
no difference. The data always move in the direction of the most 
entropic group, except for very small values of diff.
}
\description{
Takes as input the conditions vector and returns a matrix of
scale values of the same dimension as the number of samples
and the number of mc.samples used by the aldex() or aldex.clr()
function.
}
\examples{

# conditions is a vector describing the data
data(selex)
# subset for efficiency
conds <- c(rep("NS", 7), rep("S", 7))
mu.vec <- aldex.makeScaleMatrix(1, 0.2, conds, mc.samples=128)

}
\references{
Please use the citation given by \code{citation(package="ALDEx")}.
}
\seealso{
\code{\link{aldex.clr}}, \code{\link{aldex}}
}
\author{
Greg Gloor, Michelle Pistner Nixon
}
